Change code to accept decimal numbers
Since a hashmap cannot accept a fractional number, a struct was made to
contain the integral and fractional part of the decimal number. Methods
were added to create a new FloatNum struct and to convert the struct to
a string. The code was also changed to accomadate the new struct.